+++ packages/browser/test/manual/test-code.js
@@ -62,3 +62,9 @@ Sentry.withScope(scope => {
   Sentry.captureException(new TypeError('bar'));
   Sentry.captureMessage('baz');
 });
+
+var xhr = new XMLHttpRequest();
+xhr.onload = () => console.log('loaded'); // This throws error
+// xhr.addEventListener("load", () => console.log('loaded')); This does not throw error
+xhr.open('GET', 'https://httpbin.org/get');
+xhr.send();