+++ CHANGELOG.md
@@ -3,6 +3,9 @@
 ## Unreleased
 
 - [browser] ref: Mangle more stuff, reduce bundle size
+- [integrations] feat: Add tracing integration
+- [hub] feat: Add tracing related function to scope and hub (`Scope.startSpan`, `Scope.setSpanContext`, `Hub.traceHeaders`)
+- [hub] feat: Add new function to Scope `setContext`
 
 ## 5.1.1