+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 - [integrations] feat: Add tracing integration
 - [hub] feat: Add tracing related function to scope and hub (`Scope.startSpan`, `Scope.setSpanContext`, `Hub.traceHeaders`)
 - [hub] feat: Add new function to Scope `setContext`
+<<<<<<< HEAD
 - [hub] feat: Add new function to Scope `setTransaction`
+=======
+- [node] fix: Expose lastEventId method
+>>>>>>> c03f26a3b58d3a61272715087a4c4a370837ac96
 
 ## 5.1.1