+++ CHANGELOG.md
@@ -2,11 +2,13 @@
 
 ## Unreleased
 
-- [browser] ref: Mangle more stuff, reduce bundle size
 - [integrations] feat: Add tracing integration
 - [hub] feat: Add tracing related function to scope and hub (`Scope.startSpan`, `Scope.setSpanContext`, `Hub.traceHeaders`)
 - [hub] feat: Add new function to Scope `setContext`
 - [hub] feat: Add new function to Scope `setTransaction`
+- [browser] fix: Fixed a bug if Sentry was initialized multiple times: Fix #2043
+- [browser] ref: Mangle more stuff, reduce bundle size
+- [browser] fix: Support for ram bundle frames
 - [node] fix: Expose lastEventId method
 
 ## 5.1.1