+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## 5.1.2
 
 - [browser] fix: Fixed a bug if Sentry was initialized multiple times: Fix #2043
 - [browser] ref: Mangle more stuff, reduce bundle size